refactor: ImageBox 구조 변경
