+++ client/src/components/HomePage/PostImageBox/styles.tsx
@@ -1,0 +1,8 @@
+import styled, { css } from 'styled-components';
+
+const FlexWrap = styled.div`
+  display: flex;
+  align-items: center;
+`;
+
+export { FlexWrap };
